+++ collection-service/__brick__/{{collectionName.snakeCase()}}_service.dart
@@ -1,22 +1,39 @@
-class {{collectionName.pascalCase()}}Service extends TurboCollectionService<{{collectionObjectName.pascalCase()}}, {{collectionName.pascalCase()}}Api> {
-  {{collectionName.pascalCase()}}Service() : super(api: {{collectionName.pascalCase()}}Api.locate);
+import 'package:get_it/get_it.dart';
+import 'package:turbo_firestore_api/turbo_firestore_api.dart';
+
+class {{collectionName.pascalCase()}}Service extends TCollectionService<{{collectionObjectName.pascalCase()}}Dto, {{collectionObjectName.pascalCase()}}> {
+  {{collectionName.pascalCase()}}Service({
+    required super.collection,
+    required super.modelBuilder,
+    super.apiBuilder,
+    super.defaultValue,
+    super.initialValue,
+    super.streamBuilder,
+    super.initialiseStream,
+    super.firestoreCacheService,
+    super.initialSortFilteredListsMap,
+    super.modelDocsBuilder,
+  });
 
   // 📍 LOCATOR ------------------------------------------------------------------------------- \\
 
   static {{collectionName.pascalCase()}}Service get locate => GetIt.I.get();
-  static void registerLazySingleton() => GetIt.I.registerLazySingleton({{collectionName.pascalCase()}}Service.new);
+  static void registerLazySingleton() => GetIt.I.registerLazySingleton(
+    () => {{collectionName.pascalCase()}}Service(
+      collection: ,
+      modelBuilder: ,
+      initialiseStream: ,
+    ),
+    dispose: (param) async => await param.dispose(),
+  );
 
   // 🧩 DEPENDENCIES -------------------------------------------------------------------------- \\
   // 🎬 INIT & DISPOSE ------------------------------------------------------------------------ \\
   // 👂 LISTENERS ----------------------------------------------------------------------------- \\
   // ⚡️ OVERRIDES ----------------------------------------------------------------------------- \\
-
-  @override
-  FutureOr<Stream<List<{{collectionObjectName.pascalCase()}}>>> Function(User user) get stream => throw UnimplementedError();
-
-// 🎩 STATE --------------------------------------------------------------------------------- \\
-// 🛠 UTIL ---------------------------------------------------------------------------------- \\
-// 🧲 FETCHERS ------------------------------------------------------------------------------ \\
-// 🏗️ HELPERS ------------------------------------------------------------------------------- \\
-// 🪄 MUTATORS ------------------------------------------------------------------------------ \\
+  // 🎩 STATE --------------------------------------------------------------------------------- \\
+  // 🛠 UTIL ---------------------------------------------------------------------------------- \\
+  // 🧲 FETCHERS ------------------------------------------------------------------------------ \\
+  // 🏗️ HELPERS ------------------------------------------------------------------------------- \\
+  // 🪄 MUTATORS ------------------------------------------------------------------------------ \\
 }