+++ lib/router/home_router.dart
@@ -1,0 +1,35 @@
+import 'package:get_it/get_it.dart';
+import 'package:turbo_modules/routing/abstracts/t_route.dart';
+import 'package:turbo_modules/routing/abstracts/t_router.dart';
+import 'package:turbo_modules/routing/services/app_routing_service.dart';
+import 'package:turbo_modules/routing/services/navigation_tab_service.dart';
+
+class HomeRouter extends TRouter {
+  HomeRouter({
+    required super.initialRoute,
+    required super.appRoutingService,
+    required super.navigationTabService,
+  });
+
+  // 📍 LOCATOR ------------------------------------------------------------------------------- \\
+
+  static HomeRouter Function() get lazyLocate => () => GetIt.I.get();
+  static HomeRouter get locate => GetIt.I.get();
+  static void registerFactory({required TRoute initialRoute}) => GetIt.I.registerFactory(
+    () => HomeRouter(
+      initialRoute: initialRoute,
+      appRoutingService: AppRoutingService.locate,
+      navigationTabService: NavigationTabService.locate,
+    ),
+  );
+
+  // 🧩 DEPENDENCIES -------------------------------------------------------------------------- \\
+  // 🎬 INIT & DISPOSE ------------------------------------------------------------------------ \\
+  // 👂 LISTENERS ----------------------------------------------------------------------------- \\
+  // ⚡️ OVERRIDES ----------------------------------------------------------------------------- \\
+  // 🎩 STATE --------------------------------------------------------------------------------- \\
+  // 🛠 UTIL ---------------------------------------------------------------------------------- \\
+  // 🧲 FETCHERS ------------------------------------------------------------------------------ \\
+  // 🏗️ HELPERS ------------------------------------------------------------------------------- \\
+  // 🪄 MUTATORS ------------------------------------------------------------------------------ \\
+}