+++ scripts/__brick__/scripts/dart_fix.sh
@@ -4,7 +4,7 @@
 cd ../lib || exit
 
 # Run flutter build with delete-conflicting-outputs
-flutter clean && flutter pub get && flutter pub run build_runner build --delete-conflicting-outputs
+flutter clean && flutter pub get && dart run build_runner build --delete-conflicting-outputs
 
 # Find and format files excluding specific patterns
 find . -name "*.dart" \