+++ view/__brick__/{{viewName.snakeCase()}}_arguments.dart
@@ -1,17 +1,20 @@
-import 'package:{{packageName.snakeCase()}}/core/abstracts/view_arguments.dart';
-import 'package:{{packageName.snakeCase()}}/core/constants/k_keys.dart';
+import 'package:json_annotation/json_annotation.dart';
+import 'package:turbo_modules/routing/abstracts/i_view_arguments.dart';
 
-class {{viewName.pascalCase()}}Arguments extends ViewArguments {
-{{viewName.pascalCase()}}Arguments({required this.id});
+part '{{viewName.snakeCase()}}_arguments.g.dart';
 
-  final String? id;
+@JsonSerializable(
+  includeIfNull: true,
+  explicitToJson: true,
+)
+class {{viewName.pascalCase()}}Arguments extends IViewArguments {
+  {{viewName.pascalCase()}}Arguments({
+    required this.origin,
+  });
 
+  final {{viewName.pascalCase()}}Origin origin;
+
+  factory {{viewName.pascalCase()}}Arguments.fromJson(Map<String, dynamic> json) => _${{viewName.pascalCase()}}ArgumentsFromJson(json);
   @override
-  Map<String, dynamic> toMap() => {
-      kKeysId: id,
-    };
-
-  factory {{viewName.pascalCase()}}Arguments.fromMap(Map<String, dynamic> map) => {{viewName.pascalCase()}}Arguments(
-        id: map[kKeysId] as String,
-      );
+  Map<String, dynamic> toJson() => _${{viewName.pascalCase()}}ArgumentsToJson(this);
 }