+++ view/__brick__/{{viewName.snakeCase()}}_origin.dart
@@ -1,3 +1,0 @@
-enum {{viewName.pascalCase()}}Origin {
-  core,
-}